Config based ftrace analysis framework
Updated the existing script to work on a config to collect data from
the Linux ftrace. The tool works in two steps. First, based on the
config collects the needed table and creates a table in a sqlite
database. Second, runs the queries that are in the config, which are
sql statements.

Signed-off-by: Santosh Sivaraj <santosiv@in.ibm.com>