Updated config to contain multiple trace profiles
Now the trace profile which needs to be taken can be given using the
option '-t'/'--type'. This defaults to ftrace now. Selecting a
different profile has no effect now, since the regexps also should
ideally come from the config or generated at run time. The regexp
is hardcoded for a ftrace line as of now.

Signed-off-by: Santosh Sivaraj <santosiv@in.ibm.com>